Fix EntrySize property of StringEntry and extract StringEntry and StringEntryList into their own files
